src/requirements.txt
pandas
sqlalchemy
python-dotenv
psycopg2_binary
yfinance